+++ src/main/java/com/keyrus/virtualStore/saleOrder/SaleOrderServiceImpl.java
@@ -3,6 +3,8 @@ package com.keyrus.virtualStore.saleOrder;
 import com.keyrus.virtualStore.customer.CustomerModel;
 import com.keyrus.virtualStore.customer.ICustomerRepository;
 import com.keyrus.virtualStore.exception.VirtualStoreException;
+import com.keyrus.virtualStore.product.IProductService;
+import com.keyrus.virtualStore.product.ProductModel;
 import com.keyrus.virtualStore.saleOrderProduct.ISaleOrderProductRepository;
 import com.keyrus.virtualStore.saleOrderProduct.SaleOrderProductDTO;
 import com.keyrus.virtualStore.saleOrderProduct.SaleOrderProductModel;
@@ -24,10 +26,13 @@ public class SaleOrderServiceImpl implements  ISaleOrderService{
     @Autowired
     private ICustomerRepository customerRepository;
 
+    @Autowired
+    private IProductService productService;
+
     @Override
     public void addSaleOrder(SaleOrderDTO saleOrder) throws VirtualStoreException {
-        try {
 
+        try {
             SaleOrderModel saleOrderModel = new SaleOrderModel();
             saleOrderModel.setSaleOrderDate(saleOrder.getSaleOrderDate());
             String email = saleOrder.getCustomerOrder().getEmail();
@@ -35,20 +40,30 @@ public class SaleOrderServiceImpl implements  ISaleOrderService{
             if(customer == null){
                 throw new VirtualStoreException("The customer with this email doesn't exist");
             }
-
+            int newAvailableQuantity;
             saleOrderModel.setCustomerOrder(customer);
-            saleOrderModel = saleOrderRepository.saveAndFlush(saleOrderModel);
+            saleOrderModel = saleOrderRepository.save(saleOrderModel);
             List<SaleOrderProductDTO> products = saleOrder.getProducts();
             float totalPrice = 0;
 
-            for(SaleOrderProductDTO product : products){
+            for(SaleOrderProductDTO orderProductDTO : products){
+                Long productId = orderProductDTO.getProduct().getId();
+                ProductModel product = productService.findProduct(productId);
+
+                //Update the available quantity
+                newAvailableQuantity = product.getAvailableQuantity() - orderProductDTO.getQuantity();
+                if(newAvailableQuantity < 0 ){
+                    throw new VirtualStoreException("The product doesn't have the requested quantity");
+                }
+                product.setAvailableQuantity(newAvailableQuantity);
+                product = productService.updateProduct(product.getId(),product);
 
                 SaleOrderProductModel saleOrderProduct = new SaleOrderProductModel();
-                saleOrderProduct.setProduct(product.getProduct());
+                saleOrderProduct.setProduct(product);
                 saleOrderProduct.setSaleOrder(saleOrderModel);
-                saleOrderProduct.setQuantity(product.getQuantity());
-                saleOrderProductRepository.saveAndFlush(saleOrderProduct);
-                totalPrice = totalPrice + product.getQuantity()*product.getProduct().getPrice();
+                saleOrderProduct.setQuantity(orderProductDTO.getQuantity());
+                saleOrderProductRepository.save(saleOrderProduct);
+                totalPrice = totalPrice + orderProductDTO.getQuantity() * product.getPrice();
 
             }
             saleOrderModel.setTotalPrice(totalPrice);
@@ -116,6 +131,10 @@ public class SaleOrderServiceImpl implements  ISaleOrderService{
     @Override
     public void deleteSaleOrder(Long id) throws VirtualStoreException {
         try {
+            SaleOrderModel saleOrder = saleOrderRepository.findOne(id);
+            if(saleOrder == null){
+                throw new VirtualStoreException("The sale order doesn't exist");
+            }
             saleOrderRepository.delete(id);
         } catch (HibernateJdbcException e) {
             throw new VirtualStoreException("This operation is unavailable right now. Try later");
